refactor: extract state schema and status helpers
